Size now is printed
size is printed underneath the file name

+++ FM/Filer.Data/Models/mFile.cs
@@ -10,7 +10,7 @@ namespace Filer.Data.Models
     {
        
         public string Name { get; set; }
-        public long Size { get; set; }
+        public double Size { get; set; }
         public int Dir_Id { get; set; }
        
         public Byte[] bytes { get; set; }
